Refactor ActivityProvider to QueryProvider and update related components for improved search functionality
