+++ sportsphere-app/components/ActivityCardList.jsx
@@ -5,14 +5,41 @@ import { SPACING } from '../global'
 import { collection, onSnapshot } from "firebase/firestore";
 import { db } from "../Firebase/firebaseSetup";
 import { useEffect, useState, useContext } from 'react';
-import { useActivity, ActivityContext } from "../context/ActivityProvider";
+import { QueryContext } from "../context/QueryProvider";
+import { readAllFiles } from '../Firebase/firebaseHelper';
+
 
 export default function ActivityCardList() {
-  const {activityItems, setActivityItems} = useContext(ActivityContext);
+  const {searchQuery} = useContext(QueryContext);
+
+  const collectionName = "activities";
+  const [activityItems, setActivityItems] = useState([]);
+
+  function handleActivityItems(newItems) {
+    setActivityItems(newItems);
+  }
+
+  useEffect(() => {
+    // const unsubscribe = onSnapshot(collection(db, collectionName), (querySnapshot) => {
+    //   const currActivityItems = [];
+    readAllFiles(collectionName, handleActivityItems, (error) => {
+      console.log("Error fetching activities", error.message);
+    });
+  }, []);
+
+  const filteredActivityItems = activityItems.filter(item => {
+    const terms = searchQuery.toLowerCase().split(' ');
+    return terms.some(term =>
+      item.activityName.toLowerCase().includes(term) ||
+      item.venue.toLowerCase().includes(term) ||
+      item.description.toLowerCase().includes(term) ||
+      item.date.toLowerCase().includes(term)
+    );
+  });
 
   return (
     <FlatList
-      data={activityItems}
+      data={filteredActivityItems}
       keyExtractor={(item) => item.id}
       renderItem={({ item }) => (
         <ActivityCard
