+++ sportsphere-app/utils/readDBHelper.js
